[owc] FileManager now has the "mod" prefix; modifying a resource (used for tools). FileManager now has FileAccess requirement, to ensure safety for file writing/reading and querying. FileInfo struct is now used to identify a file. Added a function to validate a path with FileAccess, as well as a bool if the asset dir can be written to. Added a foreachFile function that allows looping through all files in a directory, as well as getFile which returns FileInfo. foreachFileRecurse can be used for showing all files recursively. [cmake] Added folders for sorting [tools] Now use the FileManager; instead of using a batch file and a simple conversion script. Fixed linking issues. [app] Added a simple test for showing files in app's directories. [ogc] Now links to glslang correctly. Now uses spirv_cross only in spvhelper, to avoid linking issues.

+++ ogc/src/graphics/format/oishcompiler.cpp
@@ -101,7 +101,7 @@ bool oiSH::compileSource(ShaderSource &source, bool useFile) {
 
 	glslang::InitializeProcess();
 
-	u32 len = useFile ? source.files.size() : source.src.size();
+	u32 len = (u32)(useFile ? source.files.size() : source.src.size());
 	auto it = source.src.begin();
 
 	for (u32 i = 0; i < len; ++i){
